Moved test3.js -> test.js
